requirements.txt
astroid==2.4.2
autopep8==1.5.4
certifi==2020.6.20
chardet==3.0.4
idna==2.10
isort==5.6.4
lazy-object-proxy==1.4.3
mccabe==0.6.1
Pillow==8.0.0
psutil==5.7.2
pycodestyle==2.6.0
pylint==2.6.0
reportlab==3.5.53
requests==2.24.0
six==1.15.0
toml==0.10.1
urllib3==1.25.10
wrapt==1.12.1
